feat divide expo and native notifications
