+++ src/screens/SettingScreen.tsx
@@ -3,7 +3,7 @@ import { useStackNavigation } from '../services/route'
 import { useSafeArea } from 'react-native-safe-area-context'
 import { useAuthState, useAuthActions } from '../store/hooks'
 import { useStyles, MakeStyles, useColors } from '../services/design'
-import { usePushNotifications } from '../services/secure'
+import { usePushNotifications } from '../services/notifications/notifications'
 import { TouchableOpacity, ScrollView, Text, StyleSheet, View } from 'react-native'
 import { ShadowBase } from '../components/atoms'
 import { Header } from '../components/organisms'
