+++ src/services/secure.ts
@@ -2,14 +2,11 @@ import { useEffect, useState, useCallback } from 'react'
 import { InteractionManager } from 'react-native'
 import { db } from '../repositories/firebase'
 import { setCertificateImage } from '../repositories/certificate'
-import { updateSecure, getSecure } from '../repositories/secure'
-import Notifications from 'expo-notifications'
+import { updateSecure } from '../repositories/secure'
 import * as ImagePicker from 'expo-image-picker'
 import * as Permissions from 'expo-permissions'
-import Constants from 'expo-constants'
 import { ImageInfo } from 'expo-image-picker/build/ImagePicker.types'
 import { Secure, buildSecure } from '../entities'
-import _ from 'lodash'
 
 const usersRef = db.collection('users')
 const getSecureRef = (uid: string) => {
@@ -89,79 +86,4 @@ export const useCertificateEditTools = (uid: string) => {
   }, [uid, uploadCertificateURL])
 
   return { currentCertificateURL, uploadCertificateURL, onChangeUpdateCertificateURL, upload }
-}
-
-export const usePushNotifications = (uid: string) => {
-  const [deviceToken, setDeviceToken] = useState<string | null>(null)
-
-  const getTokenWithAsk = useCallback(async () => {
-    const permissionResponse = await Permissions.getAsync(Permissions.NOTIFICATIONS)
-    let finalStatus = permissionResponse.status
-
-    if (permissionResponse.status !== 'granted') {
-      const askPermissionResponse = await Permissions.askAsync(Permissions.NOTIFICATIONS)
-      finalStatus = askPermissionResponse.status
-    }
-
-    if (finalStatus !== 'granted') {
-      return null
-    }
-
-    const token = await Notifications.getExpoPushTokenAsync()
-    return token
-  }, [])
-
-  // MEMO: effect内では、Permissionの許可申請を行わない。
-  useEffect(() => {
-    const getToken = async () => {
-      try {
-        const response = await Permissions.getAsync(Permissions.NOTIFICATIONS)
-        if (response.status !== 'granted') return
-
-        const pushNotification = await Notifications.getExpoPushTokenAsync()
-        const token = pushNotification.data
-        if (!token) return
-
-        setDeviceToken(token)
-      } catch (e) {
-        console.warn(e)
-      }
-    }
-    getToken()
-  }, [])
-
-  const onAccept = useCallback(async () => {
-    if (!Constants.isDevice) {
-      return alert('エミュレーターでは、プッシュ通知を許可できません。')
-    }
-
-    let updateToken = deviceToken
-    if (!updateToken) {
-      const pushNotification = await getTokenWithAsk()
-      const token = pushNotification.data
-      if (!token) return
-      updateToken = token
-    }
-
-    const currentSecure = await getSecure(uid)
-    const secure: Secure = {
-      pushTokens: !_.isEmpty(currentSecure.pushTokens)
-        ? _.uniq([...currentSecure.pushTokens, updateToken])
-        : [updateToken]
-    }
-    const { result } = await updateSecure(uid, secure)
-    return { result }
-  }, [deviceToken, getTokenWithAsk, uid])
-
-  const onReject = useCallback(async () => {
-    const currentSecure = await getSecure(uid)
-    const secure: Secure = {
-      pushTokens: !_.isEmpty(currentSecure.pushTokens) ? _.pull(currentSecure.pushTokens, deviceToken) : []
-    }
-    const { result } = await updateSecure(uid, secure)
-
-    return { result }
-  }, [deviceToken, uid])
-
-  return { onAccept, onReject, deviceToken }
 }
